+++ frontend/src/features/game/hooks/useCanvasSize.ts
@@ -1,12 +1,18 @@
-import { useWindowSize } from './useWindowSize';
+import { useLayoutEffect, useMemo } from 'react';
 
-export const useCanvasSize = (): {
-  canvasWidth: number;
-  canvasHeight: number;
-} => {
-  const size = useWindowSize();
-  const canvasWidth = size[0] - 300;
-  const canvasHeight = (size[0] - 300) / 2;
+export const useCanvasSize = (): number[] => {
+  const size = useMemo(() => [0, 0], []);
+  useLayoutEffect(() => {
+    const updateSize = (): void => {
+      size[0] = window.innerWidth - 300;
+      size[1] = window.innerHeight - 300;
+    };
 
-  return { canvasWidth, canvasHeight };
+    window.addEventListener('resize', updateSize);
+    updateSize();
+
+    return () => window.removeEventListener('resize', updateSize);
+  }, []);
+
+  return size;
 };
